Add an endpoint for returning Chef package data.
This introduces a `/api/v1/metadata/chef` endpoint which provides Chef
package data. Chef changed the way that packages can be determined, but
now `Mixlib::Install` allows us to fetch this in Ruby.

Instead of reverse engineering something that might well change in
future, this uses `Mixlib::Install` to fetch the result and returns it
as JSON. The main boxes project can than (hopefully) reliably fetch from
this self-hosted endpoint.

* Handles parameter validation using `sinatra-param`.
* Sets the content type and formatting correctly using `Sinatra::JSON`
  from `sinatra-contrib`.
* Tests a successful response, but also one which might fail.
* Adds `VCR` to record the hilariously huge requests that are made.